Successfully extract point
Refactoring logic currently not working but fine for now
